+++ iduka/database/migrations/2020_10_13_150128_create_cabangs_table.php
@@ -19,8 +19,8 @@ class CreateCabangsTable extends Migration
             $table->string('alamat_cabang');
             $table->string('jumlah_karyawan');
             $table->string('gambar_karyawan');
-            $table->bigInteger('umkm_id')->unsigned();
-            $table->bigInteger('user_id')->unsigned();
+            $table->integer('umkm_id')->unsigned();
+            $table->integer('user_id')->unsigned();
 
         });
