+++ iduka/database/migrations/2020_10_13_152217_create_produk_favorits_table.php
@@ -16,7 +16,7 @@ class CreateProdukFavoritsTable extends Migration
         Schema::create('produk_favorits', function (Blueprint $table) {
             $table->bigIncrements('produk_favorit_id');
             $table->bigInteger('konsumen_id')->unsigned();
-            $table->bigInteger('produk_id')->unsigned();
+            $table->integer('produk_id')->unsigned();
 
         });
