+++ iduka/database/migrations/2020_10_13_152916_create_pendaftaran_umkms_table.php
@@ -17,8 +17,8 @@ class CreatePendaftaranUmkmsTable extends Migration
             $table->bigIncrements('pendaftaran_umkm_id');
             $table->string('status_pendaftaran');
             $table->string('dokumen_pendukung');
-            $table->bigInteger('umkm_id')->unsigned();
-            $table->bigInteger('pengelola_id')->unsigned();
+            $table->integer('umkm_id')->unsigned();
+            $table->integer('pengelola_id')->unsigned();
             $table->timestamps('tanggal_pendaftaran');
 
         });
