+++ iduka/database/migrations/2020_10_14_091625_create_transaksi_kasir_details_table.php
@@ -15,8 +15,8 @@ class CreateTransaksiKasirDetailsTable extends Migration
     {
         Schema::create('transaksi_kasir_details', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('produk_id')->unsigned();
-            $table->bigInteger('transaksi_kasir_id')->unsigned();
+            $table->integer('produk_id')->unsigned();
+            $table->integer('transaksi_kasir_id')->unsigned();
             $table->timestamps();
         });
